Simplify request interface
